TPT3-92: tests are refactored
