--- legacy-data-access-services/src/test/java/gov/ca/cwds/cms/data/access/service/impl/HappyPathHelper.java
@@ -1,52 +1,0 @@
-package gov.ca.cwds.cms.data.access.service.impl;
-
-import gov.ca.cwds.cms.data.access.Constants.StaffPersonPrivileges;
-import gov.ca.cwds.cms.data.access.dto.ClientEntityAwareDTO;
-import gov.ca.cwds.cms.data.access.dto.PlacementHomeEntityAwareDTO;
-import gov.ca.cwds.data.legacy.cms.entity.Client;
-import gov.ca.cwds.data.legacy.cms.entity.PlacementHome;
-import gov.ca.cwds.data.legacy.cms.entity.SubstituteCareProvider;
-import gov.ca.cwds.security.realm.PerryAccount;
-import java.util.HashSet;
-import java.util.Set;
-
-/**
- * @author CWDS CALS API Team
- */
-
-public final class HappyPathHelper {
-
-  private HappyPathHelper() {}
-
-  public static PlacementHomeEntityAwareDTO prepareSuccessfulPlacementHomeEntityAwareDTO() {
-    PlacementHomeEntityAwareDTO placementHomeEntityAwareDTO = new PlacementHomeEntityAwareDTO();
-    PlacementHome placementHome = new PlacementHome();
-    placementHome.setAgeToNo((short) 23);
-    placementHome.setZipNo("11111");
-    placementHome.setLaPZipno("11111");
-    placementHome.setpZipNo("11111");
-    SubstituteCareProvider substituteCareProvider = new SubstituteCareProvider();
-    substituteCareProvider.setZipNo("11111");
-    placementHome.setPrimarySubstituteCareProvider(substituteCareProvider);
-    placementHomeEntityAwareDTO.setEntity(placementHome);
-    return placementHomeEntityAwareDTO;
-  }
-
-  public static PerryAccount getPrincipal(String privilege) {
-    PerryAccount perryAccount = new PerryAccount();
-    Set<String> privileges = new HashSet<>();
-    privileges.add(privilege);
-    perryAccount.setPrivileges(privileges);
-    return perryAccount;
-  }
-
-  public static PerryAccount getPlacementFacilityHappyPathPrincipal() {
-    return getPrincipal(StaffPersonPrivileges.USR_PRV_RESOURCE_MGMT_PLACEMENT_FACILITY_MAINT);
-  }
-
-  public static ClientEntityAwareDTO clientEntityAwareDTO(Client client) {
-    ClientEntityAwareDTO clientEntityAwareDTO = new ClientEntityAwareDTO();
-    clientEntityAwareDTO.setEntity(client);
-    return clientEntityAwareDTO;
-  }
-}
