+++ legacy-data-access-services/src/test/java/gov/ca/cwds/cms/data/access/service/impl/R03232Test.java
@@ -5,25 +5,25 @@ import org.junit.Test;
 /**
  * Created by TPT2 on 12/21/2017.
  */
-public class R03232Test extends BaseDocToolRulesTest {
+public class R03232Test extends BaseDocToolRulesPlacementHomeTest {
   
   //zip no
   @Test
   public void testZipNoShort() {
     placementHomeEntityAwareDTO.getEntity().setpZipNo("1");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
   @Test
   public void testZipNoLong() {
     placementHomeEntityAwareDTO.getEntity().setpZipNo("11111111");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
   @Test
   public void testZipNoFormat() {
     placementHomeEntityAwareDTO.getEntity().setpZipNo("aa");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
 
@@ -32,19 +32,19 @@ public class R03232Test extends BaseDocToolRulesTest {
   @Test
   public void testpZipNoShort() {
     placementHomeEntityAwareDTO.getEntity().setpZipNo("1");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
   @Test
   public void testpZipNoLong() {
     placementHomeEntityAwareDTO.getEntity().setpZipNo("11111111");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
   @Test
   public void testpZipNoFormat() {
     placementHomeEntityAwareDTO.getEntity().setpZipNo("aa");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
   //La P zip
@@ -52,19 +52,19 @@ public class R03232Test extends BaseDocToolRulesTest {
   @Test
   public void testLaPZipNoShort() {
     placementHomeEntityAwareDTO.getEntity().setLaPZipno("1");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
   @Test
   public void testLaPZipNoLong() {
     placementHomeEntityAwareDTO.getEntity().setLaPZipno("11111111");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
   @Test
   public void testLaPZipNoFormat() {
     placementHomeEntityAwareDTO.getEntity().setLaPZipno("aa");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
   //SCP zip no
@@ -72,19 +72,19 @@ public class R03232Test extends BaseDocToolRulesTest {
   @Test
   public void testSCPZipNoShort() {
     placementHomeEntityAwareDTO.getEntity().getPrimarySubstituteCareProvider().setZipNo("1");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
   @Test
   public void testSCPZipNoLong() {
     placementHomeEntityAwareDTO.getEntity().getPrimarySubstituteCareProvider().setZipNo("11111111");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
   @Test
   public void testSCPZipNoFormat() {
     placementHomeEntityAwareDTO.getEntity().getPrimarySubstituteCareProvider().setZipNo("aaa");
-    checkPlacementHomeService("R-03232");
+    check("R-03232");
   }
 
 }
