+++ legacy-data-access-services/src/test/java/gov/ca/cwds/cms/data/access/service/impl/R03855Test.java
@@ -10,7 +10,7 @@ import org.junit.Test;
 /**
  * Created by TPT2 on 12/11/2017.
  */
-public class R03855Test extends BaseDocToolRulesTest {
+public class R03855Test extends BaseDocToolRulesPlacementHomeTest {
 
   @Test
   public void testMaxAgeInvalid() throws Exception {
